Se borraron archivos repetidos y se cambiaron nombres a los que se quedaron. Se modificó el footer para cuadrar dimensiones e íconos. Se agregaron las páginas de estilos y se modificaron archivos para coincidir con rutas.
