+++ src/App.jsx
@@ -1,5 +1,7 @@
 import React from 'react'
-import Home from './pages/Home'
+import Home from './pages/HomePage'
+import AboutUs from './pages/AboutUs'
+import Login from './pages/LoginPage'
 
 function App() {
   return (
