+++ src/main.jsx
@@ -1,14 +1,14 @@
 import { StrictMode } from 'react'
 import { createRoot } from 'react-dom/client'
-import Footer from './components/Footer.jsx'
 import { BrowserRouter } from 'react-router-dom';
+import './index.css'
+import App from './App.jsx';
 
 
 createRoot(document.getElementById('root')).render(
   <StrictMode>
   <BrowserRouter>
-
-    <Footer/> 
+    <App/> 
   </BrowserRouter>
   </StrictMode>,
 )
