--- src/pages/Home.jsx
@@ -1,10 +1,0 @@
-import React from 'react'
-import Footer from '../components/Footer'
-
-function Home() {
-  return (
-    <Footer/>
-  )
-}
-
-export default Home
