+++ src/pages/HomePage.jsx
@@ -1,0 +1,10 @@
+import React from 'react'
+import Footer from '../components/Footer'
+
+function Home() {
+  return (
+    <Footer/>
+  )
+}
+
+export default Home
